Sign Up View
Sign up button can create accounts, you should be able to log in  from
the website in a browser to confirm that they work. The cancel button
also changes the screen to the log in page. There is a crash when you
try to create accounts with taken usernames. I'll try to solve this
issue later.